reorganizing model diffusion filters
